GEODE-6109: Reintroduce Build job to precheckin pipeline

+++ ci/scripts/execute_build.sh
@@ -36,36 +36,43 @@ fi
 
 ROOT_DIR=$(pwd)
 BUILD_DATE=$(date +%s)
-EMAIL_SUBJECT="results/subject"
-EMAIL_BODY="results/body"
 
-GEODE_BUILD_VERSION_FILE=${ROOT_DIR}/geode-build-version/number
-GEODE_RESULTS_VERSION_FILE=${ROOT_DIR}/results/number
-GEODE_BUILD_VERSION_NUMBER=$(grep "versionNumber *=" geode/gradle.properties | awk -F "=" '{print $2}' | tr -d ' ')
-GEODE_BUILD_DIR=/tmp/geode-build
-GEODE_PULL_REQUEST_ID_FILE=${ROOT_DIR}/geode/.git/id
+# Precheckin does not get a geode-build-version
+if [ -e "${ROOT_DIR}/geode-build-version" ] ; then
+  GEODE_BUILD_VERSION_FILE=${ROOT_DIR}/geode-build-version/number
+  GEODE_RESULTS_VERSION_FILE=${ROOT_DIR}/results/number
+  GEODE_BUILD_VERSION_NUMBER=$(grep "versionNumber *=" geode/gradle.properties | awk -F "=" '{print $2}' | tr -d ' ')
+  GEODE_BUILD_DIR=/tmp/geode-build
+  GEODE_PULL_REQUEST_ID_FILE=${ROOT_DIR}/geode/.git/id
 
-if [ -e "${GEODE_PULL_REQUEST_ID_FILE}" ]; then
-  GEODE_PULL_REQUEST_ID=$(cat ${GEODE_PULL_REQUEST_ID_FILE})
-  FULL_PRODUCT_VERSION="geode-pr-${GEODE_PULL_REQUEST_ID}"
-else
-  CONCOURSE_VERSION=$(cat ${GEODE_BUILD_VERSION_FILE})
-  CONCOURSE_PRODUCT_VERSION=${CONCOURSE_VERSION%%-*}
-  GEODE_PRODUCT_VERSION=${GEODE_BUILD_VERSION_NUMBER}
-  CONCOURSE_BUILD_SLUG=${CONCOURSE_VERSION##*-}
-  BUILD_ID=${CONCOURSE_VERSION##*.}
-  FULL_PRODUCT_VERSION=${GEODE_PRODUCT_VERSION}-${CONCOURSE_BUILD_SLUG}
-  echo "Concourse VERSION is ${CONCOURSE_VERSION}"
-  echo "Geode product VERSION is ${GEODE_PRODUCT_VERSION}"
-  echo "Build ID is ${BUILD_ID}"
+  if [ -e "${GEODE_PULL_REQUEST_ID_FILE}" ]; then
+    GEODE_PULL_REQUEST_ID=$(cat ${GEODE_PULL_REQUEST_ID_FILE})
+    FULL_PRODUCT_VERSION="geode-pr-${GEODE_PULL_REQUEST_ID}"
+  else
+    CONCOURSE_VERSION=$(cat ${GEODE_BUILD_VERSION_FILE})
+    CONCOURSE_PRODUCT_VERSION=${CONCOURSE_VERSION%%-*}
+    GEODE_PRODUCT_VERSION=${GEODE_BUILD_VERSION_NUMBER}
+    CONCOURSE_BUILD_SLUG=${CONCOURSE_VERSION##*-}
+    BUILD_ID=${CONCOURSE_VERSION##*.}
+    FULL_PRODUCT_VERSION=${GEODE_PRODUCT_VERSION}-${CONCOURSE_BUILD_SLUG}
+    echo "Concourse VERSION is ${CONCOURSE_VERSION}"
+    echo "Geode product VERSION is ${GEODE_PRODUCT_VERSION}"
+    echo "Build ID is ${BUILD_ID}"
+  fi
+
+  echo -n "${FULL_PRODUCT_VERSION}" > ${GEODE_RESULTS_VERSION_FILE}
 fi
 
-echo -n "${FULL_PRODUCT_VERSION}" > ${GEODE_RESULTS_VERSION_FILE}
-
-DEFAULT_GRADLE_TASK_OPTIONS="--parallel --console=plain --no-daemon"
+if [[ ${PARALLEL_GRADLE:-"true"} == "true" ]]; then
+  PARALLEL_GRADLE="--parallel"
+else
+  PARALLEL_GRADLE=""
+fi
+DEFAULT_GRADLE_TASK_OPTIONS="${PARALLEL_GRADLE} --console=plain --no-daemon"
+GRADLE_SKIP_TASK_OPTIONS=""
 
 SSHKEY_FILE="instance-data/sshkey"
-SSH_OPTIONS="-i ${SSHKEY_FILE} -o ConnectionAttempts=60 -o StrictHostKeyChecking=no"
+SSH_OPTIONS="-i ${SSHKEY_FILE} -o ConnectionAttempts=60 -o StrictHostKeyChecking=no -o ServerAliveInterval=60"
 
 INSTANCE_IP_ADDRESS="$(cat instance-data/instance-ip-address)"
 
@@ -88,10 +95,12 @@ if [ -v CALL_STACK_TIMEOUT ]; then
   ssh ${SSH_OPTIONS} geode@${INSTANCE_IP_ADDRESS} "tmux new-session -d -s callstacks; tmux send-keys  ~/capture-call-stacks.sh\ ${PARALLEL_DUNIT}\ ${CALL_STACK_TIMEOUT} C-m"
 fi
 
-GRADLE_COMMAND="./gradlew \
+GRADLE_ARGS="\
     ${DEFAULT_GRADLE_TASK_OPTIONS} \
+    ${GRADLE_SKIP_TASK_OPTIONS} \
     ${GRADLE_GLOBAL_ARGS} \
-    build install -x test"
+    build install javadoc spotlessCheck rat checkPom -x test"
 
-echo "${GRADLE_COMMAND}"
-ssh ${SSH_OPTIONS} geode@${INSTANCE_IP_ADDRESS} "mkdir -p tmp && cd geode && ${SET_JAVA_HOME} && ${GRADLE_COMMAND}"
+EXEC_COMMAND="mkdir -p tmp && cd geode && ${SET_JAVA_HOME} && ./gradlew ${GRADLE_ARGS}"
+echo "${EXEC_COMMAND}"
+ssh ${SSH_OPTIONS} geode@${INSTANCE_IP_ADDRESS} "${EXEC_COMMAND}"
